fix: enhance HMR for locale JSON updates to prevent full page reloads
